Changed structure, completed statistic view
