+++ app/src/main/java/pl/artsit/flexgoals/http/services/QuantitativeGoalService.java
@@ -6,8 +6,7 @@ import pl.artsit.flexgoals.http.goals.GoalAchieveCallback;
 import pl.artsit.flexgoals.http.goals.GoalGetCallback;
 import pl.artsit.flexgoals.http.goals.GoalUpdateCallback;
 import pl.artsit.flexgoals.http.goals.PredefinedGoalCallback;
-import pl.artsit.flexgoals.http.user.UserCallback;
-import pl.artsit.flexgoals.model.goal.QuantitativeGoalUpdateData;
+import pl.artsit.flexgoals.model.goal.quantitative.QuantitativeGoalUpdateData;
 import pl.artsit.flexgoals.model.goal.quantitative.PredefinedQuantitativeGoal;
 import pl.artsit.flexgoals.model.goal.quantitative.QuantitativeGoal;
 import pl.artsit.flexgoals.model.goal.quantitative.QuantitativeGoalData;
