+++ app/src/main/java/pl/artsit/flexgoals/model/goal/quantitative/QuantitativeGoalUpdateData.java
@@ -1,4 +1,4 @@
-package pl.artsit.flexgoals.model.goal;
+package pl.artsit.flexgoals.model.goal.quantitative;
 
 public class QuantitativeGoalUpdateData {
     private Integer id;
