+++ app/src/main/java/pl/artsit/flexgoals/ui/goals/GoalQuantitativeFragment.java
@@ -1,4 +1,4 @@
-package pl.artsit.flexgoals.ui.addGoals;
+package pl.artsit.flexgoals.ui.goals;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
