+++ app/src/main/res/layout/fragment_goal_final.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.addGoals.GoalFinalFragment">
+    tools:context=".ui.goals.GoalFinalFragment"
+    android:background="@drawable/background">
 
     <TextView
         android:id="@+id/textViewAddGoalsTitle"
