+++ app/src/main/res/layout/fragment_goal_preview.xml
@@ -6,7 +6,7 @@ android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:layout_marginBottom="10dp"
 android:id="@+id/nav_add_goal"
-tools:context=".ui.addGoals.GoalQuantitativePreviewFragment"
+tools:context=".ui.goals.GoalQuantitativePreviewFragment"
 android:background="@drawable/background">
 
     <ScrollView
