+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -43,19 +43,19 @@
 
     <fragment
         android:id="@+id/nav_quantitative_goals"
-        android:name="pl.artsit.flexgoals.ui.addGoals.GoalQuantitativeFragment"
+        android:name="pl.artsit.flexgoals.ui.goals.GoalQuantitativeFragment"
         android:label="Quantitative Goals "
         tools:layout="@layout/fragment_goal_quantitative" />
 
     <fragment
         android:id="@+id/nav_final_goals"
-        android:name="pl.artsit.flexgoals.ui.addGoals.GoalFinalFragment"
+        android:name="pl.artsit.flexgoals.ui.goals.GoalFinalFragment"
         android:label="Final Goals "
         tools:layout="@layout/fragment_goal_final" />
 
     <fragment
         android:id="@+id/nav_goal_preview"
-        android:name="pl.artsit.flexgoals.ui.addGoals.GoalQuantitativePreviewFragment"
+        android:name="pl.artsit.flexgoals.ui.goals.GoalQuantitativePreviewFragment"
         android:label="Postep"
         tools:layout="@layout/fragment_add_goals" />
